pyproject.toml
[tool.poetry]
name = "testfastapi"
version = "0.1.0"
description = ""
authors = ["Daniel-Astaptscev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
